+++ test/unit/compose/volume.spec.ts
@@ -1,12 +1,7 @@
 import { expect } from 'chai';
-import { SinonStub, stub } from 'sinon';
 import Volume from '~/src/compose/volume';
-import * as logTypes from '~/lib/log-types';
-import * as logger from '~/src/logger';
 
-import { createVolume, withMockerode } from '~/test-lib/mockerode';
-
-describe('compose/volume', () => {
+describe('compose/volume: unit tests', () => {
 	describe('creating a volume from a compose object', () => {
 		it('should use proper defaults when no compose configuration is provided', () => {
 			const volume = Volume.fromComposeObject(
@@ -167,96 +162,6 @@ describe('compose/volume', () => {
 		});
 	});
 
-	describe('creating a docker volume from options', () => {
-		before(() => {
-			stub(logger, 'logSystemEvent');
-		});
-
-		afterEach(() => {
-			(logger.logSystemEvent as SinonStub).reset();
-		});
-
-		after(() => {
-			(logger.logSystemEvent as SinonStub).restore();
-		});
-
-		it('should use defaults to create the volume when no options are given', async () => {
-			await withMockerode(async (mockerode) => {
-				const volume = Volume.fromComposeObject(
-					'one_volume',
-					1032480,
-					'deadbeef',
-				);
-
-				await volume.create();
-
-				expect(mockerode.createVolume).to.have.been.calledOnceWith({
-					Name: '1032480_one_volume',
-					Driver: 'local',
-					Labels: {
-						'io.balena.supervised': 'true',
-						'io.balena.app-uuid': 'deadbeef',
-					},
-					DriverOpts: {},
-				});
-			});
-		});
-
-		it('should pass configuration options to the engine', async () => {
-			await withMockerode(async (mockerode) => {
-				const volume = Volume.fromComposeObject(
-					'one_volume',
-					1032480,
-					'deadbeef',
-					{
-						driver_opts: {
-							opt1: 'test',
-						},
-						labels: {
-							'my-label': 'test-label',
-						},
-					},
-				);
-
-				await volume.create();
-
-				expect(mockerode.createVolume).to.have.been.calledOnceWith({
-					Name: '1032480_one_volume',
-					Driver: 'local',
-					Labels: {
-						'my-label': 'test-label',
-						'io.balena.supervised': 'true',
-						'io.balena.app-uuid': 'deadbeef',
-					},
-					DriverOpts: {
-						opt1: 'test',
-					},
-				});
-
-				expect(logger.logSystemEvent).to.have.been.calledOnceWith(
-					logTypes.createVolume,
-				);
-			});
-		});
-
-		it('should log successful volume creation to the cloud', async () => {
-			await withMockerode(async (mockerode) => {
-				const volume = Volume.fromComposeObject(
-					'one_volume',
-					1032480,
-					'deadbeef',
-				);
-
-				await volume.create();
-
-				expect(mockerode.createVolume).to.have.been.calledOnce;
-				expect(logger.logSystemEvent).to.have.been.calledOnceWith(
-					logTypes.createVolume,
-				);
-			});
-		});
-	});
-
 	describe('comparing volume configuration', () => {
 		it('should ignore name and supervisor labels in the comparison', () => {
 			expect(
@@ -407,128 +312,6 @@ describe('compose/volume', () => {
 					}),
 				),
 			).to.be.true;
-		});
-	});
-
-	describe('removing volumes', () => {
-		before(() => {
-			stub(logger, 'logSystemEvent');
-		});
-
-		afterEach(() => {
-			(logger.logSystemEvent as SinonStub).reset();
-		});
-
-		after(() => {
-			(logger.logSystemEvent as SinonStub).restore();
-		});
-
-		it('should remove the volume from the engine if it exists', async () => {
-			const dockerVolume = createVolume({
-				Name: '1234_aaa',
-			});
-
-			await withMockerode(
-				async (mockerode) => {
-					const volume = Volume.fromComposeObject('aaa', 1234, 'deadbeef');
-
-					// Check engine state before (this is really to test that mockerode is doing its job)
-					expect((await mockerode.listVolumes()).Volumes).to.have.lengthOf(1);
-					expect(await mockerode.getVolume('1234_aaa').inspect()).to.deep.equal(
-						dockerVolume.inspectInfo,
-					);
-
-					// Remove the volume
-					await volume.remove();
-
-					// Check that the remove method was called
-					expect(mockerode.removeVolume).to.have.been.calledOnceWith(
-						'1234_aaa',
-					);
-				},
-				{ volumes: [dockerVolume] },
-			);
-		});
-
-		it('should report the volume removal as a system event', async () => {
-			const dockerVolume = createVolume({
-				Name: '1234_aaa',
-			});
-
-			await withMockerode(
-				async (mockerode) => {
-					const volume = Volume.fromComposeObject('aaa', 1234, 'deadbeef');
-
-					// Check engine state before
-					expect((await mockerode.listVolumes()).Volumes).to.have.lengthOf(1);
-
-					// Remove the volume
-					await volume.remove();
-
-					// Check that the remove method was called
-					expect(mockerode.removeVolume).to.have.been.calledOnceWith(
-						'1234_aaa',
-					);
-
-					// Check that log entry was generated
-					expect(logger.logSystemEvent).to.have.been.calledOnceWith(
-						logTypes.removeVolume,
-					);
-				},
-				{ volumes: [dockerVolume] },
-			);
-		});
-
-		it('should report an error if the volume does not exist', async () => {
-			const dockerVolume = createVolume({
-				Name: '4567_bbb',
-			});
-			await withMockerode(
-				async (mockerode) => {
-					const volume = Volume.fromComposeObject('aaa', 1234, 'deadbeef');
-
-					// Check engine state before
-					expect((await mockerode.listVolumes()).Volumes).to.have.lengthOf(1);
-
-					// Remove the volume, this should not throw
-					await expect(volume.remove()).to.not.be.rejected;
-
-					// Check that the remove method was called
-					expect(mockerode.removeVolume).to.not.have.been.called;
-
-					// Check that log entry was generated
-					expect(logger.logSystemEvent).to.have.been.calledWith(
-						logTypes.removeVolumeError,
-					);
-				},
-				{ volumes: [dockerVolume] },
-			);
-		});
-
-		it('should report an error if a problem happens while removing the volume', async () => {
-			const dockerVolume = createVolume({
-				Name: '1234_aaa',
-			});
-			await withMockerode(
-				async (mockerode) => {
-					const volume = Volume.fromComposeObject('aaa', 1234, 'deadbeef');
-
-					// Stub the mockerode method to fail
-					mockerode.removeVolume.rejects('Something bad happened');
-
-					// Check engine state before
-					expect((await mockerode.listVolumes()).Volumes).to.have.lengthOf(1);
-
-					// Remove the volume, this should not throw
-					await expect(volume.remove()).to.not.be.rejected;
-
-					// Check that log entry was generated
-					expect(logger.logSystemEvent).to.have.been.calledWith(
-						logTypes.removeVolumeError,
-					);
-				},
-				{ volumes: [dockerVolume] },
-			);
 		});
 	});
 });
